perf(rooms): memoize occupant grouping in OccupantPanel
OccupantPanel re-grouped all occupants (O(n log n) sort + bare-JID grouping) on every render via an IIFE. The panel re-renders ~1:1 with its parent (room message activity, menu/connection state), so the grouping ran even when the occupants Map was unchanged.

Extract grouping into a pure, tested module (utils/occupantGrouping.ts) and memoize it on the occupants Map ref; also memoize the offline-member and hidden-ignored derivations. The pre-existing OccupantGrouping.test.ts now imports the real helper instead of a copied implementation.

Regression guard in OccupantPanel.memo.test.tsx: a re-render with unchanged occupants must not re-invoke groupOccupantsByRole (was 2 calls, now 1).
